requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.0.1
django-cron==0.6.0
django-crontab==0.7.1
djangorestframework==3.15.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
lxml==5.2.1
ntscraper==0.3.13
numpy==1.24.4
packaging==24.0
pandas==2.0.3
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
